infer structure
