+++ packages/adapters/package.json
@@ -27,7 +27,6 @@
   "homepage": "https://mitvis.github.io/olli/",
   "dependencies": {
     "@observablehq/plot": "^0.5.2",
-    "dayjs": "^1.11.7",
     "ts-loader": "^9.3.1",
     "vega": "^5.23.0",
     "vega-lite": "^5.3.0",
