+++ packages/core/package.json
@@ -27,6 +27,7 @@
   "homepage": "https://mitvis.github.io/olli/",
   "dependencies": {
     "d3-scale": "^4.0.2",
+    "dayjs": "^1.11.7",
     "vega-lite": "^5.9.3",
     "vega-statistics": "^1.9.0",
     "webpack": "^5.76.0"
